add the hyracks connector to startCluster script
git-svn-id: https://hyracks.googlecode.com/svn/branches/fullstack_genomix@2709 123451ca-8445-de46-9d55-352943316053

+++ genomix/src/main/resources/scripts/stopcc.sh
@@ -4,7 +4,7 @@ hostname
 #Kill process
 PID=`ps -ef|grep ${USER}|grep java|grep hyracks|awk '{print $2}'`
 echo $PID
-kill -9 $PID
+[ "$PID" != "" ] && kill -9 $PID
 
 #Clean up CC temp dir
 rm -rf $CCTMP_DIR/*
